Switch to event based ansiterm api
We've updated ansiterm to emit events to a channel instead of using
callbacks.

Ansiterm emits events for every type that used to be supported via
callbacks. In addition, it emits generic events for parser states that
it does not have explicit handlers for. This lets consumers handle
those events without requiring updates to the parser code.

This will allow us to handle terminfo specifics in the future.