Make reading CursorPos safer
We now only read our cursor position at startup to determine our line
number. There is no other time that we require this.

We try to be more robust in reading the cursor position as there might
be buffered events before the response is sent. We now try to handle
that gracefully.